requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\DELL\Downloads\apis\app.py: 4
Flask == 1.0.2

# C:\Users\DELL\Downloads\apis\app.py: 5
Flask_SQLAlchemy == 2.5.1

# C:\Users\DELL\Downloads\apis\app.py: 3
Keras == 2.4.3

# C:\Users\DELL\Downloads\apis\app.py: 11
PyJWT == 1.7.1

# C:\Users\DELL\Downloads\apis\app.py: 10
Werkzeug == 1.0.1

# C:\Users\DELL\Downloads\apis\app.py: 7
# C:\Users\DELL\Downloads\apis\utils.py: 5
numpy == 1.19.5

# C:\Users\DELL\Downloads\apis\app.py: 12
# C:\Users\DELL\Downloads\apis\utils.py: 4
pandas == 1.0.4

# C:\Users\DELL\Downloads\apis\app.py: 9
# C:\Users\DELL\Downloads\apis\utils.py: 7
scikit_learn == 0.22.2.post1

# C:\Users\DELL\Downloads\apis\app.py: 8
# C:\Users\DELL\Downloads\apis\utils.py: 6
scipy == 1.4.1

# C:\Users\DELL\Downloads\apis\app.py: 13
tensorflow == 2.4.1
